pylister/requirements.txt
python-frontmatter
python-slugify
